# perf(sql): speed up SQL COPY and REST API CSV import