--- a/core/src/main/java/io/questdb/cutlass/text/AbstractTextLexer.java
+++ b/core/src/main/java/io/questdb/cutlass/text/AbstractTextLexer.java
@@ -114,14 +114,10 @@ public abstract class AbstractTextLexer implements Closeable, Mutable {
             while (ptr < hi) {
                 if (!eol && !delayedOutQuote && ptr < hi - 7) {
                     long word = Unsafe.getUnsafe().getLong(ptr);
-                    long xor1 = word ^ MASK_NEW_LINE;
-                    long xor2 = word ^ MASK_CR;
-                    long xor3 = word ^ MASK_QUOTE;
-                    long xor4 = word ^ getDelimiterMask();
-                    long zeroBytesWord = SwarUtils.checkZeroByte(xor1)
-                            | SwarUtils.checkZeroByte(xor2)
-                            | SwarUtils.checkZeroByte(xor3)
-                            | SwarUtils.checkZeroByte(xor4);
+                    long zeroBytesWord = SwarUtils.checkZeroByte(word ^ MASK_NEW_LINE)
+                            | SwarUtils.checkZeroByte(word ^ MASK_CR)
+                            | SwarUtils.checkZeroByte(word ^ MASK_QUOTE)
+                            | SwarUtils.checkZeroByte(word ^ getDelimiterMask());
                     if (zeroBytesWord == 0) {
                         ptr += 7;
                         this.fieldHi += 7;
@@ -324,14 +320,10 @@ public abstract class AbstractTextLexer implements Closeable, Mutable {
             while (ptr < hi) {
                 if (!eol && !rollBufferUnusable && !useLineRollBuf && !delayedOutQuote && ptr < hi - 7) {
                     long word = Unsafe.getUnsafe().getLong(ptr);
-                    long xor1 = word ^ MASK_NEW_LINE;
-                    long xor2 = word ^ MASK_CR;
-                    long xor3 = word ^ MASK_QUOTE;
-                    long xor4 = word ^ getDelimiterMask();
-                    long zeroBytesWord = SwarUtils.checkZeroByte(xor1)
-                            | SwarUtils.checkZeroByte(xor2)
-                            | SwarUtils.checkZeroByte(xor3)
-                            | SwarUtils.checkZeroByte(xor4);
+                    long zeroBytesWord = SwarUtils.checkZeroByte(word ^ MASK_NEW_LINE)
+                            | SwarUtils.checkZeroByte(word ^ MASK_CR)
+                            | SwarUtils.checkZeroByte(word ^ MASK_QUOTE)
+                            | SwarUtils.checkZeroByte(word ^ getDelimiterMask());
                     if (zeroBytesWord == 0) {
                         ptr += 7;
                         this.fieldHi += 7;