--- a/core/src/main/java/io/questdb/cutlass/text/AbstractTextLexer.java
+++ b/core/src/main/java/io/questdb/cutlass/text/AbstractTextLexer.java
@@ -114,16 +114,16 @@ public abstract class AbstractTextLexer implements Closeable, Mutable {
             while (ptr < hi) {
                 if (!eol && !delayedOutQuote && ptr < hi - 7) {
                     long word = Unsafe.getUnsafe().getLong(ptr);
-                    long zeroBytesWord = SwarUtils.checkZeroByte(word ^ MASK_NEW_LINE)
-                            | SwarUtils.checkZeroByte(word ^ MASK_CR)
-                            | SwarUtils.checkZeroByte(word ^ MASK_QUOTE)
-                            | SwarUtils.checkZeroByte(word ^ getDelimiterMask());
+                    long zeroBytesWord = SwarUtils.markZeroBytes(word ^ MASK_NEW_LINE)
+                            | SwarUtils.markZeroBytes(word ^ MASK_CR)
+                            | SwarUtils.markZeroBytes(word ^ MASK_QUOTE)
+                            | SwarUtils.markZeroBytes(word ^ getDelimiterMask());
                     if (zeroBytesWord == 0) {
                         ptr += 7;
                         this.fieldHi += 7;
                         continue;
                     } else {
-                        long idx = SwarUtils.indexOfFirstNonZeroByte(zeroBytesWord);
+                        long idx = SwarUtils.indexOfFirstMarkedByte(zeroBytesWord);
                         ptr += idx;
                         this.fieldHi += idx;
                     }
@@ -320,25 +320,25 @@ public abstract class AbstractTextLexer implements Closeable, Mutable {
             while (ptr < hi) {
                 if (!eol && !rollBufferUnusable && !useLineRollBuf && !delayedOutQuote && ptr < hi - 7) {
                     long word = Unsafe.getUnsafe().getLong(ptr);
-                    long zeroBytesWord = SwarUtils.checkZeroByte(word ^ MASK_NEW_LINE)
-                            | SwarUtils.checkZeroByte(word ^ MASK_CR)
-                            | SwarUtils.checkZeroByte(word ^ MASK_QUOTE)
-                            | SwarUtils.checkZeroByte(word ^ getDelimiterMask());
+                    long zeroBytesWord = SwarUtils.markZeroBytes(word ^ MASK_NEW_LINE)
+                            | SwarUtils.markZeroBytes(word ^ MASK_CR)
+                            | SwarUtils.markZeroBytes(word ^ MASK_QUOTE)
+                            | SwarUtils.markZeroBytes(word ^ getDelimiterMask());
                     if (zeroBytesWord == 0) {
                         ptr += 7;
                         this.fieldHi += 7;
                         continue;
                     } else {
-                        long idx = SwarUtils.indexOfFirstNonZeroByte(zeroBytesWord);
+                        long idx = SwarUtils.indexOfFirstMarkedByte(zeroBytesWord);
                         ptr += idx;
                         this.fieldHi += idx;
                     }
                 }
 
-                final byte c = Unsafe.getUnsafe().getByte(ptr++);
+                final byte b = Unsafe.getUnsafe().getByte(ptr++);
 
-                if (checkState(ptr, c)) {
-                    doSwitch(lo, ptr, c);
+                if (checkState(ptr, b)) {
+                    doSwitch(lo, ptr, b);
                 }
             }
         } catch (LineLimitException ignore) {