Results for SQL COPY in clickbench (optimized schema with varchar), c6a.metal:

Before:
```
2024-04-05T17:00:05.272708Z I i.q.c.t.ParallelCsvFileImporter import complete [importId=085572760f310fd8, file=`/home/ubuntu/ClickBench/questdb/hits.csv`', time=153s]
```

After:
```
2024-04-05T19:33:29.855367Z I i.q.c.t.ParallelCsvFileImporter import complete [importId=6678835cfb20a947, file=`/home/ubuntu/ClickBench/questdb/hits.csv`', time=127s]
```

### Comparison for all stages

Before:
```
2024-04-05T16:57:31.915697Z I i.q.c.t.ParallelCsvFileImporter finished [importId=085572760f310fd8, phase=analyze_file_structure, file=`/home/ubuntu/ClickBench/questdb/hits.csv`, duration=0s, errors=0]
2024-04-05T16:57:35.823696Z I i.q.c.t.ParallelCsvFileImporter finished [importId=085572760f310fd8, phase=boundary_check, file=`/home/ubuntu/ClickBench/questdb/hits.csv`, duration=3s, errors=0]
2024-04-05T16:57:52.984694Z I i.q.c.t.ParallelCsvFileImporter finished [importId=085572760f310fd8, phase=indexing, file=`/home/ubuntu/ClickBench/questdb/hits.csv`, duration=17s, errors=0]
2024-04-05T17:00:02.362526Z I i.q.c.t.ParallelCsvFileImporter finished [importId=085572760f310fd8, phase=partition_import, file=`/home/ubuntu/ClickBench/questdb/hits.csv`, duration=129s, errors=0]
2024-04-05T17:00:02.507203Z I i.q.c.t.ParallelCsvFileImporter finished [importId=085572760f310fd8, phase=symbol_table_merge, file=`/home/ubuntu/ClickBench/questdb/hits.csv`, duration=0s, errors=0]
2024-04-05T17:00:04.502289Z I i.q.c.t.ParallelCsvFileImporter finished [importId=085572760f310fd8, phase=update_symbol_keys, file=`/home/ubuntu/ClickBench/questdb/hits.csv`, duration=1s, errors=0]
2024-04-05T17:00:04.502480Z I i.q.c.t.ParallelCsvFileImporter finished [importId=085572760f310fd8, phase=build_symbol_index, file=`/home/ubuntu/ClickBench/questdb/hits.csv`, duration=0s, errors=0]
2024-04-05T17:00:04.503360Z I i.q.c.t.ParallelCsvFileImporter finished [importId=085572760f310fd8, phase=move_partitions, file=`/home/ubuntu/ClickBench/questdb/hits.csv`, duration=0s, errors=0]
2024-04-05T17:00:04.634925Z I i.q.c.t.ParallelCsvFileImporter finished [importId=085572760f310fd8, phase=attach_partitions, file=`/home/ubuntu/ClickBench/questdb/hits.csv`, duration=0s, errors=0]
```

After:
```
2024-04-05T19:31:22.151939Z I i.q.c.t.ParallelCsvFileImporter finished [importId=6678835cfb20a947, phase=analyze_file_structure, file=`/home/ubuntu/ClickBench/questdb/hits.csv`, duration=0s, errors=0]
2024-04-05T19:31:25.201965Z I i.q.c.t.ParallelCsvFileImporter finished [importId=6678835cfb20a947, phase=boundary_check, file=`/home/ubuntu/ClickBench/questdb/hits.csv`, duration=3s, errors=0]
2024-04-05T19:31:33.850889Z I i.q.c.t.ParallelCsvFileImporter finished [importId=6678835cfb20a947, phase=indexing, file=`/home/ubuntu/ClickBench/questdb/hits.csv`, duration=8s, errors=0]
2024-04-05T19:33:27.542270Z I i.q.c.t.ParallelCsvFileImporter finished [importId=6678835cfb20a947, phase=partition_import, file=`/home/ubuntu/ClickBench/questdb/hits.csv`, duration=113s, errors=0]
2024-04-05T19:33:27.634467Z I i.q.c.t.ParallelCsvFileImporter finished [importId=6678835cfb20a947, phase=symbol_table_merge, file=`/home/ubuntu/ClickBench/questdb/hits.csv`, duration=0s, errors=0]
2024-04-05T19:33:29.195984Z I i.q.c.t.ParallelCsvFileImporter finished [importId=6678835cfb20a947, phase=update_symbol_keys, file=`/home/ubuntu/ClickBench/questdb/hits.csv`, duration=1s, errors=0]
2024-04-05T19:33:29.196209Z I i.q.c.t.ParallelCsvFileImporter finished [importId=6678835cfb20a947, phase=build_symbol_index, file=`/home/ubuntu/ClickBench/questdb/hits.csv`, duration=0s, errors=0]
2024-04-05T19:33:29.197285Z I i.q.c.t.ParallelCsvFileImporter finished [importId=6678835cfb20a947, phase=move_partitions, file=`/home/ubuntu/ClickBench/questdb/hits.csv`, duration=0s, errors=0]
2024-04-05T19:33:29.222566Z I i.q.c.t.ParallelCsvFileImporter finished [importId=6678835cfb20a947, phase=attach_partitions, file=`/home/ubuntu/ClickBench/questdb/hits.csv`, duration=0s, errors=0]
```

### Microbenchmarks

Before:
```
Benchmark                        Mode  Cnt       Score       Error  Units
CsvTextLexerBenchmark.testParse  avgt    3  257946.332 ± 25624.637  us/op
```

After:
```
Benchmark                        Mode  Cnt       Score       Error  Units
CsvTextLexerBenchmark.testParse  avgt    3  176653.536 ± 11096.986  us/op
```