@ideoma 2024-04-05
#### [PR Coverage check]

:heart_eyes: **pass** : 97 / 97 (100.00%)




#### file detail

|   |path|covered line|new line|coverage|
|----|----|----|----|----|
|:large_blue_circle:|io/questdb/std/SwarUtils.java|3|3|100.00%|
|:large_blue_circle:|io/questdb/cutlass/text/CsvTextLexer.java|4|4|100.00%|
|:large_blue_circle:|io/questdb/cutlass/text/GenericTextLexer.java|6|6|100.00%|
|:large_blue_circle:|io/questdb/cutlass/text/CsvFileIndexer.java|31|31|100.00%|
|:large_blue_circle:|io/questdb/cutlass/text/AbstractTextLexer.java|39|39|100.00%|
|:large_blue_circle:|io/questdb/cutlass/text/CopyTask.java|14|14|100.00%|