@ideoma 2024-04-09
#### [PR Coverage check]

:heart_eyes: **pass** : 93 / 93 (100.00%)




#### file detail

|   |path|covered line|new line|coverage|
|----|----|----|----|----|
|:large_blue_circle:|io/questdb/cutlass/line/tcp/LineTcpParser.java|6|6|100.00%|
|:large_blue_circle:|io/questdb/std/SwarUtils.java|3|3|100.00%|
|:large_blue_circle:|io/questdb/cutlass/text/CsvTextLexer.java|4|4|100.00%|
|:large_blue_circle:|io/questdb/cutlass/text/GenericTextLexer.java|6|6|100.00%|
|:large_blue_circle:|io/questdb/cutlass/text/CsvFileIndexer.java|27|27|100.00%|
|:large_blue_circle:|io/questdb/cutlass/text/AbstractTextLexer.java|34|34|100.00%|
|:large_blue_circle:|io/questdb/cutlass/text/CopyTask.java|13|13|100.00%|